Pronostico del tiempo detalle

+++ app/src/main/java/car/gov/co/carserviciociudadano/openweather/model/List.java
@@ -5,7 +5,7 @@ package car.gov.co.carserviciociudadano.openweather.model;
  */
 
 public class List {
-    public int dt;
+    public long dt;
     public Main main;
     public java.util.List<Weather> weather;
     public Clouds clouds;
